add payment gateway

+++ cart/utils/cart.py
@@ -1,5 +1,3 @@
-from decimal import Decimal
-
 from shop.models import Product
 
 CART_SESSION_ID = 'cart'
@@ -17,7 +15,7 @@ class Cart:
         for product in products:
             cart[str(product.id)]['product'] = product
         for item in cart.values():
-            item['total_price'] = Decimal(item['price']) * Decimal(item['quantity'])
+            item['total_price'] = int(item['price']) * int(item['quantity'])
             yield item
 
     def add_cart_session(self):
@@ -45,7 +43,7 @@ class Cart:
         self.session.modified = True
 
     def get_total_price(self):
-        return sum(Decimal(item['price']) * item['quantity'] for item in self.cart.values())
+        return sum(int(item['price']) * item['quantity'] for item in self.cart.values())
 
     def clear(self):
         del self.session[CART_SESSION_ID]
